Share-ovanje fajla, čitanje message-fajla

+++ EncryptedFileSystem/Controllers/FileSystemController.cs
@@ -15,8 +15,80 @@ namespace EncryptedFileSystem.Controllers
     {
         private static readonly string CERTS_PATH = Application.StartupPath + "\\Certificates";
         private static readonly string FS_PATH = Application.StartupPath + "\\FileSystem";
-        private static readonly string USERS_PATH = Application.StartupPath + "\\Users";
         private static readonly string SHARED_PATH = Application.StartupPath + "\\Shared";
+
+        public static void ShareFile(string senderName, string receiverName, FileInfo file, string algorythm, string password)
+        {
+            if (!file.Exists)
+                throw new EfsException("Greška prilikom otvaranja fajla!");
+            else
+            {
+                User sender = UserController.ReadUserInfo(senderName);
+
+                if (!Verify(sender.Name, sender.HashAlgorythm, file))
+                    throw new EfsException("Ne možete podijeliti datoteku jer je njen integritet narušen!");
+                else
+                {
+                    //TODO: posto na mom fajl sistemu sve treba biti kriptovano mojim simetricnim algoritmom, PRVO-PRVO trebam dekriptovati svoje fajlove
+
+
+                    //prvo se kriptuju originalni fajl i digitalno potpisan originalni fajl
+                    var files = EncryptForSharing(file, algorythm, password);
+
+                    //zatim u fajl sa nazivom 'posiljalac_datum_vrijemeSlanja' upisuju se redom:
+                    //naziv poslanog fajla, simetricni algoritam kojim je originalni fajl kriptovan i kljuc za enkripciju
+                    //taj fajl se kriptuje javnim kljucem primaoca tako da ce ga samo primalac moci dekriptovati
+                    string sourcePath = file.FullName.Substring(0, file.FullName.LastIndexOf("\\") + 1);
+                    string[] content = { file.Name, algorythm, password };
+                    var encryptedInfoFile = EncryptInfoFile(sourcePath, senderName, receiverName, content);
+
+                    //prethodni rezultati se smjestaju u shared folder, u folder sa nazivom primaoca
+                    string destinationPath = SHARED_PATH + "\\" + receiverName;
+                    if (!Directory.Exists(destinationPath))
+                        Directory.CreateDirectory(destinationPath);
+
+                    File.Move(encryptedInfoFile.FullName, destinationPath + "\\" + encryptedInfoFile.Name.Remove(0, 3));
+                    File.Move(files[0].FullName, destinationPath + "\\" + files[0].Name.Remove(0, 3));
+                    File.Move(files[1].FullName, destinationPath + "\\" + files[1].Name.Remove(0, 3));
+                }
+            }
+        }
+
+        private static FileInfo EncryptInfoFile(string sourcePath, string sender, string receiver, string[] content)
+        {
+            string infoFile = sourcePath + "message_" + sender + "_" + DateTime.Now.ToString().Replace(' ', '_').Replace(':', '-') + ".txt";
+            File.WriteAllLines(infoFile, content);
+
+            string outInfoFile = infoFile.Insert(infoFile.LastIndexOf("\\") + 1, "###");
+            string publicKey = CERTS_PATH + "\\public\\" + receiver + "_public.key";
+
+            string command = "openssl rsautl -encrypt -in " + infoFile + " -out " + outInfoFile + " -inkey " + publicKey + " -pubin";
+            CommandPrompt.ExecuteCommand(command);
+
+            File.Delete(infoFile);
+
+            return new FileInfo(outInfoFile);
+        }
+
+        private static List<FileInfo> EncryptForSharing(FileInfo file, string algorythm, string key)
+        {
+            string outFile = file.FullName.Insert(file.FullName.LastIndexOf("\\") + 1, "###");
+            string command = "openssl " + algorythm + " -in " + file.FullName + " -out " + outFile + " -nosalt -k " + key + " -base64";
+            CommandPrompt.ExecuteCommand(command);
+            FileInfo output1 = new FileInfo(outFile);
+
+            string inFile2 = file.FullName.Insert(file.FullName.LastIndexOf('.'), "#"); //ovo je digitalni potpis izabranog fajla, trebam i njega proslijediti
+            string outFile2 = inFile2.Insert(inFile2.LastIndexOf("\\") + 1, "###");
+            string command2 = "openssl " + algorythm + " -in " + inFile2 + " -out " + outFile2 + " -nosalt -k " + key + " -base64";
+            CommandPrompt.ExecuteCommand(command2);
+            FileInfo output2= new FileInfo(outFile2);
+
+            return new List<FileInfo>
+            {
+                output1,
+                output2
+            };
+        }
 
         public static void OpenFile(FileInfo selectedFile, string username)
         {
@@ -31,6 +103,26 @@ namespace EncryptedFileSystem.Controllers
                     System.Diagnostics.Process.Start(selectedFile.FullName);
                 }
             }
+        }
+
+        public static string ReadMessage(FileInfo infoFile, string username)
+        {
+            string content = null;
+            if (infoFile.Exists)
+            {
+                string privateKey = CERTS_PATH + "\\private\\" + username + "_private.key";
+                string tmpPath = FS_PATH + "\\tmp.txt";
+                string command = "openssl rsautl -decrypt -in " + infoFile.FullName + " -out " + tmpPath + " -inkey " + privateKey;
+                CommandPrompt.ExecuteCommand(command);
+                
+                content = File.ReadAllText(tmpPath);
+                if (string.IsNullOrEmpty(content))
+                    throw new EfsException("Nemate dozvolu za ovu operaciju!");
+
+                File.Delete(tmpPath);
+            }
+
+            return content;
         }
 
         public static void CreateNewFile(string username, string relativePath, string fileName, string content)
@@ -98,11 +190,6 @@ namespace EncryptedFileSystem.Controllers
             return hostPath;
         }
 
-        public static void DeleteFromShared()
-        {
-            throw new NotImplementedException();
-        }
-
         public static void MakeDirectory(string username, string relativePath, string name)
         {
             string absolutePath = FS_PATH + "\\" + relativePath + "\\" + name;
@@ -131,8 +218,6 @@ namespace EncryptedFileSystem.Controllers
                 throw new EfsException("Došlo je do greške prilikom upload-a. Izaberite drugu datoteku.");
             else
             {
-                //TODO: Dodati provjeru naziva, ne mogu se dodati datoteke sa istim nazivom
-
                 User user = UserController.ReadUserInfo(username);
                 FileInfo selectedFile = new FileInfo(hostPath);
 
@@ -141,11 +226,14 @@ namespace EncryptedFileSystem.Controllers
                 else
                 {
                     string newPath = efsPath + "\\" + selectedFile.Name.Replace(' ', '_').Replace('#', '_');
-                    File.Copy(hostPath, newPath);
+                    if (File.Exists(newPath))
+                        throw new EfsException("Na izabranoj putanji postoji datoteka sa istim nazivom!");
+                    else
+                    {
+                        File.Copy(hostPath, newPath);
 
-                    Digest(user.Name, user.HashAlgorythm, newPath);
-                    //File.Delete(newPath);
-
+                        Digest(user.Name, user.HashAlgorythm, newPath);
+                    }
                     //TODO: Provjeri da li još šta ovdje treba
                 }
             }                
@@ -155,9 +243,12 @@ namespace EncryptedFileSystem.Controllers
         {
             if (o.GetType() == typeof(FileInfo))
             {
-                File.Delete(path);
+                if(File.Exists(path))
+                    File.Delete(path);
+                
                 string signature = path.Insert(path.LastIndexOf('.'), "#");
-                File.Delete(signature);
+                if(File.Exists(signature))
+                    File.Delete(signature);
             }
             else
                 Directory.Delete(path, true);
